Run caddy in host mode for IP number logging.
The previous config for caddy ran within the caddy network defined
in the networks.sh script, now it is run in host network mode. This
changes the received IP numbers, as previously all logs would show
the docker node in the caddy network as the originating IP for any
requests. Now the real client address is shown in caddy's logs as
well as in X-Forwarded-For: headers send when proxying requests.

This also required changing the way the nextcloud and homeassistant
containers are referenced for reverse_proxy configurations.